requirements.txt
# Core Dependencies
python>=3.10.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Network & Packet Capture
scapy>=2.4.5
pyshark>=0.4.6
dpkt>=1.9.8
netfilterqueue>=1.1.0

# Reinforcement Learning
gym>=0.21.0
stable-baselines3>=1.7.0
tensorflow>=2.10.0
torch>=1.13.0
tensorboard>=2.10.0

# Dashboard & Visualization
dash>=2.7.0
plotly>=5.11.0
flask>=2.2.0
dash-bootstrap-components>=1.2.0

# Data Processing & Analysis
matplotlib>=3.5.0
seaborn>=0.11.0
scipy>=1.7.0

# System & Networking
psutil>=5.8.0
subprocess32>=3.5.4

# Logging & Monitoring
loguru>=0.6.0
tqdm>=4.62.0

# Configuration & Utils
pyyaml>=6.0
click>=8.0.0
python-dotenv>=0.19.0

# Optional Advanced Features
ray[rllib]>=2.0.0  # For distributed RL
wandb>=0.13.0      # For experiment tracking